requirements.txt
authlib
Django==5.2.3
django-allauth
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.32.4
sentry_sdk
virtualenv==20.26.6
whitenoise==5.2.0
azure-communication-email
django-model-utils
dj_database_url
bs4
paho-mqtt==1.6.1